--- polaris/benchmark/predictions/__init__.py
@@ -1,3 +1,0 @@
-from ._predictions import BenchmarkPredictions
-
-__all__ = ["BenchmarkPredictions"]
