+++ polaris/evaluate/utils.py
@@ -4,8 +4,7 @@ import numpy as np
 import pandas as pd
 from numpy.typing import NDArray
 
-from polaris.benchmark.predictions import BenchmarkPredictions
-from polaris.evaluate import BenchmarkResults, Metric, ResultsType
+from polaris.evaluate import BenchmarkPredictions, BenchmarkResults, Metric, ResultsType
 from polaris.utils.types import IncomingPredictionsType, PredictionsType
